Adjust radius function to be "dual" of location
The radius of a node should be inversely proportional to the "density"
of space around the node. For example, if one node is in an area of the
screen such that a pixel represents twice as much length, then the node
should have half the radius.

To enable this, the radius function should be equal to the derivative of
the space function at each point. We choose a function that approximates
this derivative.