Use arrays to represent vectors
Replace all vector math with operations on arrays, many of which mutate
their argument. This decreases object allocations by several orders of
magnitude, which allows us to run smoothly on a phone.

It is unfortunate that we lose our immutability guarantees, but this is
one place where it is worth it.